requirements.txt
# Airflow
apache-airflow==2.8.1

# Web Scraping
selenium==4.15.2

# Azure
azure-identity==1.15.0
azure-storage-blob==12.19.0
